Fix circular dependency between game-utils and game-validation
Extract BUILDING_COSTS and BUILDING_INCOME to src/lib/game-constants.ts.
Both game-utils and game-validation now import from game-constants,
breaking the circular ES module cycle that was causing the Vercel build
to fail.
